Oss release 27-0-130 created 2023-01-26-04-52
see CHANGELOG.txt for details

Original commit sha: 6b1bf2f6da2cdf589ad03722cf42cd563d9c32ba

+++ framework/Ramsh/include/Ramsh/RamshTypeInfo.h
@@ -181,7 +181,7 @@ namespace ramses_internal
     {
         inline explicit operator std::string() const
         {
-            return "int64";
+            return "int";
         }
     };
 #endif
